theta-browser/ai-agents/requirements.txt
crewai==0.63.6
crewai-tools==0.12.1
langchain==0.2.16
langchain-openai==0.1.25
langchain-anthropic==0.1.23
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
openai==1.51.2
anthropic==0.34.2
tiktoken==0.7.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
aiohttp==3.10.10
asyncio==3.4.3
uvicorn==0.31.1
fastapi==0.115.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
redis==5.1.1
celery==5.3.6
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1
mypy==1.11.2
rich==13.9.2
typer==0.12.5
pyyaml==6.0.2
beautifulsoup4==4.12.3
requests==2.32.3
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
transformers==4.45.2
torch==2.4.1
sentence-transformers==3.1.1
chromadb==0.5.11
qdrant-client==1.11.3
weaviate-client==4.8.0
pinecone-client==5.0.1
supabase==2.8.0
n8n-client==1.0.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
microsoft-graph-api==1.0.0
notion-client==2.2.1
slack-sdk==3.33.1
discord-py==2.4.0
tweepy==4.14.0
python-telegram-bot==21.7 